% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distance_models.R
\name{fit_distance_models}
\alias{fit_distance_models}
\title{Fit detection curves to formatted data}
\usage{
fit_distance_models(data, spec, maxdist = 100, bins = c(0, 10, 20, 30, 40,
  50, 75, 100), modlist = c("unif.cos1", "unif.cos12", "hn.null", "hn.cos2",
  "hn.cos23", "hr.null", "hr.poly2", "hr.poly24", "hr.poly246"))
}
\arguments{
\item{data}{Dataframe formatted for use with package \code{Distance}, e.g. output from \code{format_data}}

\item{spec}{Species name or code to subset \code{data}}

\item{maxdist}{Maximum distance (in meters) to include in analysis, i.e. truncation distance}

\item{bins}{Vector of cutpoints for binned detections. See details.}

\item{modlist}{List of detection curves to include in analysis. See details.}
}
\value{
Returns a named list including:

- The output of each model in \code{modlist}

- \code{table}, containing the table of results printed

- \code{D}, the density estimates (individuals/ha) for each stratum, calculated from the model with the lowest AIC score (of the models with include = *, see Details)

- \code{N}, the estimated total number of individuals per stratum, calculated from the same model as for \code{D} above, based on extrapolating \code{D} by the \code{Area} of each stratum specified in \code{data}. If \code{Area}=1, then \code{N} = \code{D}.
}
\description{
Fit detection curves to formatted data
}
\details{
This function will first subset a formatted point count data set (i.e. output from \code{format_data}) to a desired species and fit a standard list of detection curves to the data using detection bins. An AIC table comparing detection curve fits, including Akaike weight (w), whether fit was monotonic (mono), and the p-value from a chi-square goodness-of-fit test (chisq) will then be created and printed. Any models that are monotonic and have a chisq p-value>0.05 will be considered to have "good fit", and marked to be further considered (include = *). Of the included models, the function will plot the results of the one with the lowest AIC score and print density estimates by strata (individuals/ha). Take caution in accepting the goodness-of-fit test results, since the test weights all bins equally even though the first bins are the most important. Check the plot for good agreement between observed and predicted detections. It can be easy to pass the chi-square test with small sample sizes and poor agreement, yet difficult to pass with large sample sizes and very good agreement. If agreement is poor, or none of the models pass the chi-square test, consider lumping bins or reducing the maximum (truncation) distance.

The value for \code{bins} defaults to the cutpoints for CADC protocol VCP25, bins = c(0,10,20,30,40,50,75,100), but cutpoints can be changed for other protocols and/or dropped to lump bins together as needed. Take care in changing cutpoints so that they align with the bins originally used to collect the data, particularly when combining data using more than one protocol. The function will not check whether your bins make sense for your data!

The default list of detection functions to fit were originally derived from Thomas et al. 2010 J Appl Ecol 47:5-14. These include:
- Uniform key with cosine adjustments of order 1 (unif.cos1) or 1,2 (unif.cos12)
- Half-normal key with no adjustments (hn.null), and with cosine adjustments of order 2 (hn.cos2) or 2,3 (hn.cos23)
- Hazard-rate key with no adjustments (hr.null), and with polynomial adjustments of order 2 (hr.poly2) or 2,4 (hr.poly24)
A smaller subset of these models can be fit by specifying a vector of the names of the models to include in the \code{modlist} argument, as in: modlist=c('hn.null','hr.null). The function will pass the data and desired detection curve(s) to the \code{ds} function from package \code{Distance}, which produces copious output as each model is running. Occasionally, individual models will not fit, but those will be skipped and the other models will continue to run.
}
\examples{

\dontrun{fdat = format.data(data=dat, strata='group')}
}
\author{
Kristen Dybala, \email{kdybala@pointblue.org}
}

