% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data_flat}
\alias{format_data_flat}
\title{Format data}
\usage{
format_data_flat(data, strata = "Transect", year = NULL, strata.area = 1,
  sample = "Point", visitnum = "Visit", weights = NULL, species = "Spp",
  count = "Count", dist = NULL, dist_bin_id = "Distance.Bin.ID",
  flyover = NULL, protocol = "Protocol")
}
\arguments{
\item{data}{Original point count dataframe with individual species detections and distances in each row; ok to contain multiple species at this stage}

\item{strata}{Column name in \code{data} specifying the groups for which you want density estimates (e.g., site, vegetation age class, or a combination of the two); defaults to "Transect"; it's ok to have only 1 stratum}

\item{year}{Optional, defaults to \code{NULL}; if density estimates should be generated by year and \code{strata}, change to column name in \code{data} indicating year}

\item{strata.area}{Default of 1 produces estimates of the number of individuals per hectare; change to name of column representing actual area of each stratum (in hectares) to instead extrapolate densities and estimate total number of individuals}

\item{sample}{Column name in \code{data} specifying independent sampling locations within each stratum; defaults to "Point"}

\item{visitnum}{Column name in \code{data} defining the unique visit (survey) number or ID at each sampling location within each stratum; used to calculate sampling effort at each location; defaults to "Visit"}

\item{weights}{Optional, defaults to \code{NULL}; Column name in \code{data} to multiply sampling effort by (e.g., percent cover of target vegetation type)}

\item{species}{Column name in \code{data} containing name or code of species detected; defaults to "Spp"}

\item{count}{Column name in \code{data} containing the number of individuals for each detection; defaults to "Count"}

\item{dist}{Column name in \code{data} containing estimated distance to each species detection (in meters); defaults to \code{NULL}}

\item{dist_bin_id}{Defaults to 'Distance.Bin.ID'; currently only partially supported for protocol VCP25}

\item{flyover}{Optional, defaults to \code{NULL}; Column name in \code{data} flagging flyover detections for removal}

\item{protocol}{Column name in \code{data} containing point count protocol ID; defaults to "Protocol"; drops fixed radius protocol FR50 and produces warning if data contains more than one other protocol}
}
\value{
Returns a data frame including columns required by package \code{Distance}, including: \code{Region.Label}, \code{Area}, \code{Sample.Label}, \code{Effort}, and \code{distance}.
Also returns \code{species}, \code{count}, \code{protocol}, and any other unused columns remaining in original data.
}
\description{
Format data
}
\details{
This function is the first step in conducting a distance analysis, formatting point count data into the data structure with column names required by the package \code{Distance}.

Default values are intended for work with California Avian Data Center (CADC) outputs, using column names produced by downloading data in "Project Lead" mode, but should be adaptable to other formats.

Effort is calculated as the number of unique values in \code{visitnum} for each \code{strata} and optionally \code{year}. Effort may be further weighted by values in \code{weights}. For error checking, the function will print the total number of unique strata and sampling locations identified, as well as the range of the number of visits to each sampling location.

The function requires providing either \code{dist}, a numeric column with detection distances (which may represent the midpoint of distance bins), or \code{dist_bin_id}. If only \code{dist_bin_id} is provided, the function will attempt to translate the standard bin IDs provided by CADC to the corresponding numeric midpoint values for the protocol VCP25 only. For all other protocols, the function currently produces an error. All distances are assumed to be in meters and strata area in hectares.

Any values in \code{flyover} other than '', ' ', or \code{NA} are assumed to flag birds flying over a sampling location (but not using the sampling location), and will be removed.
}
\examples{

\dontrun{fdat = format.data(data=dat, strata='group')}
}
\author{
Kristen Dybala, \email{kdybala@pointblue.org}
}

