% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{Plot results of distance sampling}
\usage{
plot_density(obj, type = "D", year = F)
}
\arguments{
\item{obj}{Model output from \code{fit_distance_models}}

\item{type}{Type of output to plot, either \code{D} for density estimates or \code{N} for abundance estimates. Default is \code{D}.}

\item{year}{Defaults to \code{FALSE}. Change to \code{TRUE} if strata include estimates by year and you wish to plot trends.}
}
\value{
Returns a ggplot2 object for further customization.
}
\description{
Plot results of distance sampling
}
\examples{

\dontrun{
fdat = format_data(data=dat, strata='group')
results = fit_distance_models(fdat, 'speciesname', bins=c(0,10,20,30,40,50,75,100))
plot_density(results, 'D', year=T) + theme_classic() +
scale_color_manual(values=c('#682C7','#64AD34'))
}

}
\author{
Kristen Dybala, \email{kdybala@pointblue.org}
}
